+++ src/constants.ts
@@ -3,7 +3,7 @@ import { ApiCreator } from './types';
 
 export const defaultSubscriptionsPath = '/subscriptions';
 export const defaultGraphiQLPath = '/graphiql';
-export const defaultRootPath = '/';
+export const defaultEndpointPath = '/';
 export const defaultApi: AvailableApi = {};
 
 export const defaultApiCreator: ApiCreator<any, any> = () => ({});